FALA Z 30.07 DOMKNIETA: main nie przechodzil typechecku, to jest jego brakujaca polowa
Poprzedni commit ROZCIAL spojna zmiane na pol. Dowod: na HEAD `tsc --noEmit` daje
5 bledow, a wszystkie naprawia dokladnie ta fala:

  hold-release-modal.client.tsx:40  TS2307 brak modulu '../../_components/pending-quality-signoff'
  ncr-close-modal.client.tsx:136-137 TS2339 'closedAt' / 'signatureHash' nie istnieja
  ncrs/_components/labels.ts:179    TS2353 'rootCauseRequired' nie istnieje w typie
  lp-detail-actions.ts:391          TS2339 'releasedAt' nie istnieje

Dwa niezalezne dowody, ze to rozciecie, a nie nowa praca:
  - hold-release-modal.client.tsx jest SLEDZONY i NIEZMIENIONY, a importuje
    pending-quality-signoff — plik, ktorego na HEAD NIE MA
  - ncrs/_components/labels.ts jest SLEDZONY i NIEZMIENIONY, a juz podaje
    rootCauseRequired; typ przyjmujacy to pole zostal w drzewie roboczym

KOMPLET (to jest istota tego commita — 3 razy w tym repo commit bez kompletu
zepsul drzewo):
  - pending-quality-signoff.tsx (byl NIESLEDZONY, importuje go 6 plikow zrodlowych)
  - _meta/i18n-staging/{quality-ncrs,warehouse-lp,warehouse-adjustments}.json
    — to sa RUNTIME IMPORTY (qa-ncrs-labels.ts:19, lp-labels.ts:13,
    adjustments-labels.ts:21), nie komentarze. Nowe klucze istnieja tylko tutaj.

PACZKI: jakosc/dwupodpis · magazyn/dwupodpis LP + korekty · jednostki w zamowieniach
(itemUomOptions, 5 konsumentow, komplet zamkniety) · NPD pipeline + drobne.

POMIARY (suita UI, jsdom, deterministyczna):
  fala: 3545 pass / 38 fail      HEAD: 3518 pass / 44 fail
  czyli +27 przechodzacych, -6 awarii. Zero z 19 zmienionych plikow testowych
  jest czerwone. 5 czerwonych w lp-detail.test.tsx jest PRE-EXISTING (identyczne
  na HEAD, dotycza gatingu reserve/merge/split, ktorego diff nie rusza).

  Porownania suity node NIE podaje — obie serie bija w te sama mutowalna baze t3
  i wynik jest nierozstrzygalny (ten sam commit dal raz 0, raz 8 awarii).

CELOWO POZA TYM COMMITEM:
  - changeover-data.ts — join dopasowuje linie po kodzie filtrujac TYLKO po org_id,
    a unikalnosc kodu linii jest PER ZAKLAD (production_lines_org_site_code_uq).
    Odtworzone na zywo: jedno zdarzenie, dwa wiersze. Idzie osobno.
  - next-env.d.ts — autogenerowany artefakt lokalnego `next dev`
  - artefakty parity-evidence — nadpisane przebiegiem weryfikacyjnym, nie fala

NIE NAPRAWIA blokera builda: list-where-used.ts eksportuje `const` i `type`
z modulu 'use server'. `pnpm build` pada tak samo przed fala i po niej.
Osobna naprawa.

+++ apps/web/app/[locale]/(app)/(admin)/settings/npd-checklist/_actions/checklist-template-mutations.ts
@@ -262,7 +262,8 @@ export async function reorderChecklistTemplateItem(input: unknown): Promise<Muta
           where org_id = app.current_org_id()
             and template_id = $1
             and gate_code = $2
-            and sequence = $3 + $5`,
+            -- node-pg sends untyped params: bare $3 + $5 is unknown + unknown.
+            and sequence = $3::int + $5::int`,
         [parsed.data.templateId, parsed.data.gateCode, parsed.data.sequence, neighborSequence, tempOffset],
       );
